# apps/backend/cmd/server/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Arc API",
    "description": "API for Arc Consent Manager and Privacy Platform",
    "version": "1.0.0",
    "contact": {
      "name": "Arc Support",
      "email": "support@pixpivot.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Development Server"
    },
    {
      "url": "https://api.arc.pixpivot.com",
      "description": "Production Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/public/receipts/verify/{receiptNumber}": {
      "get": {
        "summary": "Verify Receipt",
        "description": "Verify the authenticity of a consent receipt.",
        "tags": ["Public API"],
        "parameters": [
          {
            "name": "receiptNumber",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Receipt details",
            "content": {
              "application/json": {
                "schema": { "type": "object" }
              }
            }
          },
          "404": { "description": "Receipt not found" }
        }
      }
    }
  }
}